Working Sunburst + json editor
